refactor(heureka): extract shared RemediationModal to eliminate triplicated modal code
